+++ tests/testthat/test-plot.R
@@ -7,18 +7,20 @@ example_data <- data.frame(
   values = c(5, 8, 7)
 )
 
-p <- plot_usmap()
+p <- plot_usmap("counties")
 q <- plot_usmap(data = statepop, values = "pop_2015")
 r <- plot_usmap(data = example_data)
+s <- plot_usmap(labels = TRUE)
 
 test_that("ggplot object is returned", {
   expect_is(p, "ggplot")
   expect_is(q, "ggplot")
   expect_is(r, "ggplot")
+  expect_is(s, "ggplot")
 })
 
 test_that("correct data is used", {
-  p_map_data <- us_map(regions = "states")
+  p_map_data <- us_map(regions = "counties")
   expect_identical(p$data, p_map_data)
 
   q_map_data <- map_with_data(statepop, values = "pop_2015")
@@ -26,6 +28,9 @@ test_that("correct data is used", {
 
   r_map_data <- map_with_data(example_data)
   expect_identical(r$data, r_map_data)
+
+  s_map_data <- us_map(regions = "states")
+  expect_identical(s$data, s_map_data)
 })
 
 test_that("layer parameters are correct", {
@@ -54,6 +59,13 @@ test_that("layer parameters are correct", {
     expect_equal(deparse(r$layers[[1]]$mapping$fill), "map_df[, values]")
     expect_equal(as.character(r$layers[[1]]$aes_params$colour), "black")
     expect_equal(r$layers[[1]]$aes_params$size, 0.4)
+
+    expect_is(s$layers[[1]], "ggproto")
+    expect_equal(deparse(s$layers[[1]]$mapping$x), "map_df$long")
+    expect_equal(deparse(s$layers[[1]]$mapping$y), "map_df$lat")
+    expect_equal(deparse(s$layers[[1]]$mapping$group), "map_df$group")
+    expect_equal(as.character(s$layers[[1]]$aes_params$colour), "black")
+    expect_equal(s$layers[[1]]$aes_params$size, 0.4)
   } else {
     expect_is(p$layers[[1]], "ggproto")
     expect_equal(deparse(p$layers[[1]]$mapping$x), "~map_df$long")
@@ -78,6 +90,13 @@ test_that("layer parameters are correct", {
     expect_equal(deparse(r$layers[[1]]$mapping$fill), "~map_df[, values]")
     expect_equal(as.character(r$layers[[1]]$aes_params$colour), "black")
     expect_equal(r$layers[[1]]$aes_params$size, 0.4)
+
+    expect_is(s$layers[[1]], "ggproto")
+    expect_equal(deparse(s$layers[[1]]$mapping$x), "~map_df$long")
+    expect_equal(deparse(s$layers[[1]]$mapping$y), "~map_df$lat")
+    expect_equal(deparse(s$layers[[1]]$mapping$group), "~map_df$group")
+    expect_equal(as.character(s$layers[[1]]$aes_params$colour), "black")
+    expect_equal(s$layers[[1]]$aes_params$size, 0.4)
   }
 })